+++ internal/resource/common/resourcelink/resource_link_schema.go
@@ -22,3 +22,20 @@ func ToSchema() map[string]schema.Attribute {
 		},
 	}
 }
+
+func SingleNestedAttribute() schema.SingleNestedAttribute {
+	return schema.SingleNestedAttribute{
+		Optional:   true,
+		Attributes: ToSchema(),
+	}
+}
+
+func CompleteSingleNestedAttribute(optional, computed, required bool, description string) schema.SingleNestedAttribute {
+	return schema.SingleNestedAttribute{
+		Optional:    optional,
+		Computed:    computed,
+		Required:    required,
+		Description: description,
+		Attributes:  ToSchema(),
+	}
+}