+++ internal/resource/config/oauth/tokenexchange/tokengeneratormapping/oauth_token_exchange_token_generator_mapping_resource.go
@@ -52,7 +52,7 @@ func (r *oauthTokenExchangeTokenGeneratorMappingResource) Schema(ctx context.Con
 	schema := schema.Schema{
 		Description: "Manages the mapping from a token exchange processor policy to a token generator.",
 		Attributes: map[string]schema.Attribute{
-			"attribute_sources":              attributesources.ToSchema(),
+			"attribute_sources":              attributesources.ToSchema(0),
 			"attribute_contract_fulfillment": attributecontractfulfillment.ToSchema(true, false),
 			"issuance_criteria":              issuancecriteria.ToSchema(),
 			"source_id": schema.StringAttribute{